Move bughunt argument types into module
